Refactor: Moved 4 task navigation files (#8544)
