+++ src/components/registration-landing-page/RegistrationLandingPage.tsx
@@ -5,20 +5,20 @@ import { useSelector } from 'react-redux';
 import { matchPath, useLocation, useNavigate, useParams } from 'react-router';
 import { useFetchRegistration } from '../../api/hooks/useFetchRegistration';
 import { useFetchRegistrationTickets } from '../../api/hooks/useFetchRegistrationTickets';
-import { ErrorBoundary } from '../../components/ErrorBoundary';
-import { PageSpinner } from '../../components/PageSpinner';
 import { ActionPanelContext } from '../../context/ActionPanelContext';
 import { LandingPageContext } from '../../context/LandingPageContext';
+import NotFound from '../../pages/errorpages/NotFound';
+import { NotPublished } from '../../pages/errorpages/NotPublished';
+import { ActionPanel } from '../../pages/public_registration/ActionPanel';
+import { PublicRegistrationContent } from '../../pages/public_registration/PublicRegistrationContent';
+import { TaskNavigation } from '../../pages/tasks/_components/TaskNavigation';
 import { RootState } from '../../redux/store';
 import { RegistrationStatus } from '../../types/registration.types';
 import { userHasAccessRight } from '../../utils/registration-helpers';
 import { doNotRedirectQueryParam, IdentifierParams, UrlPathTemplate } from '../../utils/urlPaths';
 import { hasTicketCuratorRole } from '../../utils/user-helpers';
-import NotFound from '../errorpages/NotFound';
-import { NotPublished } from '../errorpages/NotPublished';
-import { TaskNavigation } from '../tasks/_components/TaskNavigation';
-import { ActionPanel } from './ActionPanel';
-import { PublicRegistrationContent } from './PublicRegistrationContent';
+import { ErrorBoundary } from '../ErrorBoundary';
+import { PageSpinner } from '../PageSpinner';
 
 export const RegistrationLandingPage = () => {
   const user = useSelector((state: RootState) => state.user);
