+++ src/pages/tasks/dialogue/TasksDialoguePage.tsx
@@ -1,9 +1,9 @@
 import { UseQueryResult } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
 import { useOutletContext } from 'react-router';
+import { TicketList } from '../../../components/ticket-list/TicketList';
 import { TicketListDefaultValuesWrapper } from '../../../components/TicketListDefaultValuesWrapper';
 import { CustomerTicketSearchResponse } from '../../../types/publication_types/ticket.types';
-import { TicketList } from '../../messages/components/TicketList';
 
 interface TasksDialoguePageContext {
   ticketsQuery: UseQueryResult<CustomerTicketSearchResponse>;
